// generated/models/Task.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Task = {
    id: number;
    title: string;
    description: string;
    due_date: string;
    status: string;
    tags: Array<string>;
    completed_at?: string;
    created_at: string;
    updated_at: string;
};